drivers: video: Move buffer management stuffs into a separate file
Move buffer management stuffs into video_buffer.c
